Handle incorrectly escaped text (#160)
Parsing currently returns everything until the next '>' when the text
inside an element contains "<-" and "/" somewhere in the text.

"<-" is invalid text and should be "&lt;-" according to the Markup
Validation Service:

https://validator.w3.org/#validate_by_input

This commit changes detection of invalid tags to check if a tag starts
with text, as tags aren't allowed to start with hyphen or colons. Find
more information in the HTML Specification:

https://www.w3.org/TR/html/syntax.html#tag-name