Do not attempt to remove Smarty scripts
This fixes an issue where regular text between "{" and "}" is removed
from the DOM, even if it is no Smarty script. There is unfortunately
no simple way to distict scripts from regular contents. However, Smarty
scripts are server-side scripts and not supposed to be included in the
HTML output anyway.

Additionally Smarty scripts don't have to be enclosed in curly braces,
as this can be changed:

  https://www.smarty.net/docs/en/language.escaping.tpl

Since Simple HTML DOM Parser is supposed to parse HTML, Smarty scripts
can be ignored and should be removed manually by the caller if needed:

  "preg_replace("/(\{\w)(.*?)(\})/s", '');"

Note: Like before, Smarty scripts are kept in the DOM, just not replaced
by ___noise____ tags. Elements inside Smarty scripts may be interpreted
as regular HTML tags, which may break code in some situations!

References #154