Support attributes without whitespace to separate them
Parsing fails on attributes that are not separated by whitespace. For
example:

  <a href="#"title="PHP Simple HTML DOM Parser" />

Browsers silently accept the missing whitespace and continue parsing.

This commit changes the order in which the attribute name is read from
the document in order to continue parsing if a valid name was found
even if no space is put between attributes.

References #116