Don't add whitespace to current span elements
Whitespace on span elements should only be added if the span element
is part of another node. For span elements (i.e. the span element is
the current element), adding whitespace is useless.

For example:

  <p>PHP Simple <span>HTML DOM</span>Parser</p>

  Plaintext for the "p" element

  "PHP Simple HTML DOM Parser"

  Plaintext for the "span" element

  "HTML DOM "

This commit changes this behavior to add whitespace only if the span
element is a sub-node:

  Plaintext for the "span" element (using same example as above)

  "HTML DOM"