Add build script to automate releases
release.sh specifies a build script to automatically generate release
files based on the current commit tag. This is a two-stage process:

1) Tag current version: "git tag x.y", where x is the major and y the
     minor version number. For example: "git tag 1.6"

2) Build release file: "sh release.sh". For the example above, this
     will build "simplehtmldom_1_6.zip"

Note:

All ocurrences of "$Rev$" are replaced by "Rev. x.y (n)", where "x" is
the major and "y" the minor version number and "n" is the total number
of commits in the current branch.