Improve performance of charset detection
simple_html_dom::parse_charset() requires about 25% of total parsing
time when encoding is not specified by the document or custom loader
(i.e. using 'get_last_retrieve_url_contents_content_type').

This is caused by a call to $this->root->plaintext, which is the same
as calling $this->root->text(). This function returns the entire DOM
tree as text, which takes long time to generate depending on the size
of the document.

This commit changes implementation to work on the original document,
using $this->doc. It also makes more sense in this context, since
$this->root->plaintext takes the detected charset into consideration
which could result in unpredictable behavior.

testcase/slick_test.php shows a performance improvement of approx. 25%

before: 270 ms
after : 200 ms