Fix broken CSS selector attributes with value "0"
CSS selector attributes are incorrectly been recognized empty if the
value is set to "0" or "0.0". This results in "find('[id=0]')" being
recognized as "find('[id]')" or "find('#0')" being recognised as
"find('#')".

This issue is caused by using 'empty()' on values returned by the CSS
parser, which are considered empty when set to "" (empty string),
0 (0 as an integer), 0.0 (0 as a float), "0" (0 as a string), null,
false and array() (an empty array).

See also:
http://php.net/manual/en/function.empty.php#refsect1-function.empty-returnvalues

Additionally, to detect if a value is empty or not, the shorthand if-
statement "if($parameter)" was used, which behaves similar to calling
"if(empty($parameter))".

See also:
http://php.net/manual/en/language.types.boolean.php#language.types.boolean.casting

This commit replaces these conditions to explicitly check if the value
is empty (i.e. "$parameter === ''").

References #62