Remove automatic filter for tbody selectors
Selectors for 'tbody' are automatically been removed in parse_selector.
This results in the selector being ignored when specified for the find
function.

This commit removes this feature, so that 'tbody' selectors are taken
into consideration for finding (sub-)elements.

Note:

Browsers typically add the tbody element if it doesn't exist in the raw
HTML document. However, when using this parser, those elements may not
exist if parsing the raw document!

Due to this change, if the DOM doesn't contain the tbody element,
existing selectors may fail and need to be updated.

If the element exists and was specified in the selector, the result will
no longer contain header elements. Thus, indexing may break as well.

The easiest way to maintain the previous state is to remove 'tbody' from
all selectors!

References #79