Add support for attribute pipe selectors
This commit adds support for pipe selectors on attributes ("|=").

This is primarily intended to allow language subcode matches (e.g.,
the hreflang attribute on the a element in HTML).

Example: a[hreflang|="en"] matches "en-US", "en-UK", "en" but not "eng"
because the value must either match exactly or start with the specified
parameter, optionally followed by a hyphen ("-").

Note: While similar to "^=", this selector only matches attributes
where the value exactly matches or is immediately followed by a hyphen.

https://www.w3.org/TR/selectors/#attribute-selectors