Use Markdonw + MkDocs instead of custom HTML pages for the manual
MkDocs is a popular site generator for project documentation, based
on Markdown files: https://www.mkdocs.org/

Markdown is a well established markup language:
https://en.wikipedia.org/wiki/Markdown

Find more details to building site data in /manual/README.md

Note:

This commit doesn't add or change any of the exiting information
on the manual. All contents were copied from the original files
with the intention to keep everything in place as much as possible.

Chapters are placed in separate files and tabs have become chapters
in the files.