Parse entire attribute before skipping duplicates
If attributes with the same name appear twice or more on the same
element, only the first attribute should be taken into account;
dropping duplicate elements.

parse_attr currently returns control to the caller if the attribute
name already exists in $node->attr[]. This causes the parser to skip
the attribute name and continue parsing on the next character.

If the skipped attribute has a value, the next character is going to be
recognized as the beginning of another attribute. Since the attributes
list doesn't contain an attribute with that name, it is added to the
list as an attribute without value; a flag.

Example:

    <div class="h1" class="h2" />

Results in (using JSON notation):

    "div" : {
        "attr" : [
            { "name" : "class", "value" : "h1" }
            { "name" : "h2", "value" : 1 }
        ]
    }

Note that for the second attribute the name was skipped, but the value
is used for the name instead and the actual value got replaced by 1.

References #166