Remove code documentation from the source file
The updated manual now also includes all the code documentation which
makes the code documentation in the source file redundent.

Please note that the code documentation will always diverge from the
actual implementation as it is not >>tested<< in any way. This will
cause confusion at some point, which is why having the documentation
maintained "separately" and open for all users (accessible online)
will make it more likely for it being updated, because users actually
read it with the expectation to be up-to-date.