Add license and copyright statement
The source file simply states the MIT License without further details,
which is fixed by this commit.

In the root folder find the LICENSE file which includes the full MIT
license and copyright notice for this project.

Based on the information provided at
https://choosealicense.com/licenses/mit/

Note that the copyright statement must include all project authors in
order to ensure validity. For now we can assume that all contributors
simply oblige by the same license (which should be fine).

References
  Feature Request #47