Add detection for CP1251 to parse_charset
If mb_detect_encoding detects either 'CP1252' or 'ISO-8559-1' iconv
might fail because the document is actually encoded in 'CP1251' or
'ISO-8559-5' (mb_detect_encoding cannot differentiate these two).

In order to properly detect the correct encoding, use iconv with the
encoding detected by mb_detect_encoding and if it fails, use 'CP1251'
instead (in worse case it'll fail as well).

Note: This can have a slight impact on performance, but since the
document doesn't specify the encoding and this only gets executed
if support for mbstring is enabled, it should be fine.