Add the ability to remove a node recursively from the DOM tree
Currently the only way to "remove" a node from the DOM tree is by
doing the operation manually (clear outertext, remove child nodes,
etc...).

This commit adds a new function 'removeChild' to simple_html_dom_node
which takes a child node of the current element and removes it from
the DOM tree **recursively**.

This means, after the operation all nodes related to the specified
element will be deleted from the DOM tree, including whitespace.

Find a new test 'dom_test' which tests this behavior and works as
an example to show the recommended implementation.

The function name 'removeChild' was chosen from HTML DOM [1]

[1] https://www.w3schools.com/jsref/met_node_removechild.asp

References #29