Add save function to simple_html_dom_node
It is currently possible to save the entire dom to disk via a save
function in simple_html_dom, but it is not possible to do that for
a specific node.

This commit introduces a save function for nodes that works similar
to the save function for the dom.

References [feature-request:#41]