Decode HTML entities after loading the document
Currently all HTML entities are left untouched, which requires users
to decode entities manually (using html_entity_decode). That process
can be automated by the parser if the charset is known (which it is)

This commit adds a new function 'decode' to simple_html_dom to decode
all text after parsing.

Warning: This can have side-effects on some implementations that take
interest in encoded entities, because they are decoded twice.

For example

  <p>&amp;amp;</p>

is decoded by the parser to

  <p>&amp;</p>

so if it is decoded again, the result is

  <p>&</p>

References [feature-request:#52]