Fix decoding of attributes
Attributes we not included in the decoding process, which resulted
in attribute values with encoded entities.

This commit fixes this issue by decoding each attribute individually
as the last step of the decoding process.

Note that attribute values don't contain "noise", because it is
already filtered out by parse_attr(). It is therefore not necessary
to call remove_noise() on attribute values during decoding.

References [feature:#52]