Add function to remove the current node from the DOM tree
It is currently possible to remove a child node from the DOM tree
using 'removeChild($child)'. This commit adds a function to remove
the current node: 'remove()'

Note that it is not possible to remove a root node (node without
parent) as it would result in an empty DOM. In that case the DOM
can simply be cleared anyway.

References [bug:#167]