Properly strip newlines and whitespaces if $stripRN = true
load() currently removes newlines globally without concern for HTML
entities. This can cause weired behavior if the layout of the HTML
document is formatted in a non-compatible way.

Example

  <p
     class="topic"
     id="important_topic"
  >This is a very
	important topic!!!!!<
  /p>

Please note that this example exaggerates the issue. Anyway, simply
removing the newlines will result in wrong contents.

stripRN

  <p      class="topic"      id="important_topic"   >This is a very         important topic!!!!!<   /p>

Most of the whitespace has no meaning, so here is a cleaner version

  <p class="topic" id="important_topic">This is a very         important topic!!!!!</p>

While the document structure is correct, the contents changed in a
wrong way. Browsers typically replace the newline and any indentation
by a single whitespace. Resulting in the following.

  <p class="topic" id="important_topic">This is a very important topic!!!!!</p>

This commit changes the parser to take indentation into account. It
recognizes one or more tabs or spaces at the end of a line, followed
by a newline character and zeor or more tabs or spaces (indentation).

A single space character is only placed between non-HTML entities.