Use idiomatic if..elseif..else in text()
This function currently uses multiple if- and switch-satements as
guard clauses to determine what to return as text representation
(if any) for the current node.

This can be expressed in a more idiomatic way. Instead of returning
in multiple places, the conditions simply specify what to return on
which condition.