Use rtrim in text()
This function currently adds a blank line before paragraphs and a
single space after a <span> tag. Before adding the blank line for
the paragraph, however, it uses trim() to remove any whitespace.

This unfortunately also removes whitespace at the beginning of the
string, which is not responsibility of this particular line.

For <span> tags, any whitespace at the end of the current string
is left untouched, which can produce multiple whitespaces, where
only one is expected.

Using rtrim() in both cases fixes these issues.

Note that we cannot make any assumptions on the left side of the
string, because the function calls itself recursively and we have
no way of knowing if the current call is the first, or not. Thus,
applying trim() on the return value is no option.