Add optional argument $trim to $node->text()
In most cases when using

$node->text() or
$node->plaintext or
$html->plaintext

it is desireable to receive the text representation of the current
node and all descendents with no sorounding whitespace. Thus, this
function should make it possible to remove whitespace automatically.

This commit adds an optional argument $trim, which is set to true
by default in order to archive that. When $trim is set to false,
the function may return sorounding whitespace for paragraphs, spans
or nodes with whitespace in their contents.

References feature:#52