Add attribute value normalization
Currently attributes are treated as is, which means values maintain
the representation of the source document, including whitespace and
newline characters (potentially).

However, attribute values must follow the HTML specification:
- https://www.w3.org/TR/html/syntax.html#attribute-values
- https://www.w3.org/TR/xml/#AVNormalize

Based on these two specifications, attribute values are supposed to
be normalized by the parser. Especially the XML specification makes
the rules clear (see second link above).

  If the attribute type is not CDATA, then the XML processor MUST
  further process the normalized attribute value by discarding any
  leading and trailing space (#x20) characters, and by replacing
  sequences of space (#x20) characters by a single space (#x20)
  character.

This commit adds this behavior to the parser and to the find() such
that attribute values can be found and used naturally.

Please note that the $stripRN option on
- file_get_html,
- str_get_html,
- simple_html_dom->__construct() and
- simple_html_dom->load()
is used to determine if the whitespace is maintained in the DOM or
not. In either case, attribute normalization is applied to both the
parser and the find() function to harmonize selectors, no matter
the input data.

This can break compatibility on implementations that depend on
selectors with whitespace (i.e. [attr="value"] vs. [attr=" value"]).

These selectors (and variations of those) are now considered equal:

  [attr="value1 value2"]
  [attr=" value1 value2"]
  [attr="value1 value2 "]
  [attr="      value1     value2      "]
  [attr="
         value1
         value2
        "]