Fix selector parsing for attribute selectors ending on 's' or 'i'
Attributes selectors without double quotes that end on 's' or 'i'
cause a failure in the parser because 's' and 'i' at the end of an
attribute selector indicates case-sensitive (s) or case-insensitive (i)
attribute matching.

The bug is caused by the regex expression which requires **zero** or
more spaces before 's' or 'i' instead of **one** or more spaces.

References [bugs:#169]