Increase MAX_FILE_SIZE from 0.6 MB to 2.5 MiB
According to httparchive.org, the median sum of transfered bytes for
all resources requested by a webpage is about 2 MB [1]
(1.9 MB on desktop, 1.7 MB on mobile).

These statistics include all transfered bytes, including CSS, JS,
JSON, images, etc., most of which is not being used by the parser.

The median size of HTML bytes is less than 30 KB [2], which means
that the majority of data in a webpage is dead weight (for the
parser anyway).

The current MAX_FILE_SIZE limit of 0.6 MB (600000 Bytes) does not
address the increased webpage sizes from the past years and needs
to be increased.

This commit increases MAX_FILE_SIZE to 2.5 MiB, which is higher than
the median size of all elements on a webpage. Since most of the data
is not used by the parser, this should provide enough space even for
very large pages.

All data is from 2019.

[1] https://httparchive.org/reports/page-weight#bytesTotal
[2] https://httparchive.org/reports/page-weight#bytesHtml

References [feature-requests:#64]