Add simple_html_dom::expect() and simple_html_dom_node::expect()
When searching elements using the find() function, an empty array
is returned if no element was found. This makes it possible to do
a foreach loop that iterates over optional elements.

However, if the caller expects at least one element but the DOM
changed unexpectedly, the only way to make sure that at least one
element was returned is to put a check in place. These checks then
must be done at each loop.

Note, the simplest yet unconventional way to perform this check is:
`foreach($html->find('e') ?: null as $element) {...}`

In these cases the caller didn't expect the DOM to change and needs
to adjust the calling code to adapt the new DOM. By changing the
empty array into null, the code fails and an error is returned (it
is not possible to handle DOM changes inside the code anyway).

Since this requirement is very common, this commit adds a function
`expect`, which has the necessary code inside. The example above is
simplified to: `foreach($html->expect('e') as $element) {...}`.

This also clearly communicates the intent of the caller, who expects
at least one element. If there is no element, the function returns
null.

References [feature-requests:#65]