Add build script for Travis-CI
**Important**: The build script only works from the GitHub repo!
**Note**: Two lines in HtmlDocument were too long for the test to pass.

Find the current build status at:
https://travis-ci.com/simplehtmldom/simplehtmldom

Here is the documentation for Travis-CI:
https://travis-ci.com/getting_started

Further documentation to connect GitHub and Travis-CI:
https://docs.travis-ci.com/user/tutorial/

Travis-CI will automatically perform checks on the entire repository
based on the setup in the build script, whenever a commit is pushed
to GitHub.

The test involves PHP-CodeSniffer (phpcs) for code compatibility
and PHPUnit (phpunit) for unit tests.

Tests automatically take place for PHP 5.6, 7.0, 7.1, 7.2 and 7.3.

PHP 7.3 is also used to detect degrading unit tests for newer versions
of PHPUnit but don't count towards the final result (allow failure).