Add __debugInfo to make print_r and var_dump usable
Currently print_r() and in some instances var_dump() attempt to dump
a very large amount of data when being presented with the DOM.

The total memory consumption even for smaller documents will quickly
accumulate to many Gigabytes.

Example:

  $client = new HtmlWeb();
  $dom = $client->load('https://google.com/');
  print_r($dom);

**Warning** This example will require more than 16GB of memory!
**Note** I'm unable to give precise numbers because my computer ran
out of memory. The total consumption for this example could be much
higher.

The reason for this behavior is related to the fact that the DOM is
a recursive tree of nodes and the way print_r and var_dump process
them:

  print_r()

    Prints human-readable information about a variable.
    print_r [...] will also show protected and private properties of
    objects. [2]

  var_dump()

    This function displays structured information about one or more
    expressions that includes its type and value. Arrays and objects
    are explored recursively with values indented to show structure [2]

Notice that the output for both functions is not very helpful as it
includes a lot of duplication from object recursion and private data.

Fortunately both functions support the magic method __debugInfo [3],
which provides a way to customize the output for objects.

This commit adds __debugInfo to HtmlDocument and HtmlNode to provide
more useful information to users who utilize print_r and var_dump.

Example: (using the same example from above)

  $client = new HtmlWeb();
  $dom = $client->load('https://google.com/');
  print_r($dom);

With __debugInfo this creates about 300kB of data (experience may vary
because of online data).

[1] https://www.php.net/manual/en/function.print-r.php
[2] https://www.php.net/manual/en/function.var-dump.php
[3] https://www.php.net/manual/en/language.oop5.magic.php#object.debuginfo