Use magic method '__call' for lowercase method calling
simplehtmldom currently supports two calling conventions:

- medial capitals (i.e. "loadFile")
- lowercase (i.e. "load_file")

This adds unnecessary complexity not only to the library, but also
to the codebase for users of this libarary, especially when multiple
programmers with conflicting naming conventions are involved.

This commit adds the magic method '__call' [1] to 'HtmlDocument' and
'HtmlNode,' for all lowercase methods for which medial capitals are
also available. This ensures that existing code works like before.

As mentioned in the CHANGELOG, all lowercase methods are deprecated
and should be replaced by their respective medial captials version.

The lowercase functions should be removed in the next major version
of simplehtmldom.

**Note**: The reason for removing lowercase methods is because many
functions were only available as medial captials. Removing lowercase
is therefore less intrusive to existing code.

[1] https://www.php.net/manual/en/language.oop5.overloading.php#object.call