Add new Debug class
simplehtmldom requires a small amount of debugging functionality in
order to inform users about malformed documents, parsing issues and
deprecated functions.

This commit adds a new Debug class, which provides the ability to
dynamically enable and disable debugging messages and allows to set
a custom function for writing output to a custom debugger.

The intended use is like this:

  include_once 'Debug.php';
  use simplehtmldom\Debug;

  Debug::setDebugHandler('myDebugHandler'); // optional
  Debug::enable();

  // Do normal parsing here

  Debug::disable();
  Debug::setDebugHandler(); // clear debug handler

  function myDebugHandler($message) { ... }

The default debug handler is error_log [1] which means messages are
displayed in the console (CLI mode) or the error log file specified
in php.ini.

Here is an example debug message when calling 'load_file()':

  [DEBUG] [/var/www/html/simplehtmldom/HtmlDocument.php:114] [/var/
  www/html/simplehtmldom/test.php:29] "simplehtmldom\HtmlDocument->
  load_file() has been deprecated and will be removed in the next m
  ajor version of simplehtmldom. Use simplehtmldom\HtmlDocument->lo
  adFile() instead."

[1] https://www.php.net/manual/en/function.error-log.php