Fix duplicate noise restoration halts the script
This issue was introduced by b50d97b for which CDATA and Comments
store the same information twice. However, when restore_noise() is
used on either one of the elements, the other cannot be restored
because restore_noise() immediately unsets the restored element.

This document failed with the previous commit:

  <!-- <style></style> -->