Add short circuit conditions to reduce calls to copy_until_char()
The parser currently makes little to no assumptions on the current
character and always calls to copy_until_char(). This results in a
function call overhead which can be avoided in situations, where
the correct character is already in memory.

This commit adds short circuit conditions at "hotspots" in order to
reduce the number of calls to copy_until_char(). On an average HTML
document this can speed up parsing by ~5%.