Add short circuit conditions to reduce calls to copy_skip()
The parser currently makes little to no assumptions on the current
character and always calls to copy_skip(). This adds function call
overhead which can be avoided in situations where the current char
is unskippable.

This commit adds short circuit conditions at "hotspots" in order to
reduce the number of calls to copy_skip(). This can speed up parsing
up to 5-7%.