Fix HtmlDocument::createElement() and HtmlDocument::createTextNode()
createElement() and createTextNode() were both broken because they
were still calling str_get_html() which is not part of the namespace.

This commit fixes both function and makes it possible to create a
headless HtmlNode by calling 'new HtmlNode(null)'. This is necessary
to create elements without adding them to the DOM.