Fix HtmlNode::appendChild() doesn't add children to the DOM
appendChild() currently adds a node to a parent node, such that it
appears in the saved HTML, but it won't appear in search results
when using find().

This commit fixes this issue by manipulating the DOM structure. Note
that adding new elements to the DOM can slow down seek operations
considerably.