Check semantic versioning in release.sh
Packagist requires semantic versioning in order to build composer
packages automatically [1].

Although the build script is in no way related to the composer build,
it can be utilized to check if the current version number is valid for
composer.

This commit adds a check to release.sh for detection of git tags that
don't match the Semantic Versioning [2] requirement.

Note that Packagist accepts other formats as well, but simplehtmldom
should follow the "major.minor.patch-suffix" versioning
(i.e. "2.0.0-RC3" instead of "2.0-RC3")

[1] https://packagist.org/about#managing-package-versions
[2] https://semver.org/