Handle comments that are considered "malformed" by the HTML specification
Comments that are considered "malformed" by the HTML specification [1]
are incorrectly handled by the parser.

Example:

  <html>
  <!-->
  <p>Hello, World!</p>
  <!-- (Un)important details -->
  </html>

DOM tree:

  html
    comment

In fact, the HTML specification [1] strictly constraints the text that
may appear inside comments:

> Comments must have the following format:
>
> [...]
>
> 2. Optionally, text, with the additional restriction that the text
> must not start with the string ">", nor start with the string "->",
> nor contain the strings "<!--", "-->", or "--!>", nor end with the
> string "<!-".
>
> [...]

The parser currently ignores these constraints, which is why malformed
comments (like the one in the example above) break the DOM.

This commit adds additional checks when parsing comments to detect
malformed comments, inform the user (only if debugging is enabled)
and continue parsing the original document as if the comment didn't
exist.

This only applies to the first two constraints of the specification:

1) Text that starts with the string ">" (i.e. "<!-->")
2) Text that starts with the string "->" (i.e. "<!--->")

Comments may still contain the strings "<!--" or "--!>" and they may
still end with "<!-" contrary to the specification.

[1] https://html.spec.whatwg.org/multipage/syntax.html#comments