Add new class to handle HTML elements
This adds a new class `HtmlElement` with constants for all elements
defined by the HTML Standard[1] and methods to check if a given element
belongs to a certain element category.

This also replaces literal strings with constants to make it easier to
find them in the project.

The block tags and optional closing tags specified in `HtmlDocument` are
not yet covered by this.

Notice that the HTML5 standard no longer distinguishes between
block-level and inline-level elements and uses different categories
that don't match 1:1.

[1] https://html.spec.whatwg.org/multipage/