docs: Move page titles to mkdocs.yml
This removes the yml-style titles from all markdown files, which makes
them more readable in markdown viewers. This also makes it possible to
sort pages directly in mkdocs.yml. The current sort order is
alphabetically, with protected methods towards the end of the list.

This also removes the collapsible menus, which don't work right (always
scrolling to the first subtree instead of the top).

This also removes HtmlNode/is_inline_element, which no longer exists.