HtmlDocument: Check PHP_MAXPATHLEN before is_file()
This fixes an issue that can occur on some systems, where is_file()
returns a warning ('File name is longer than the maximum allowed path
length on this platform (4096)') if the specified string is longer than
PHP_MAXPATHLEN.

Notice: I am unable to reproduce this issue on my machine but found
several reports for this warning on the Internet. It appears that the
issue only occurs on some platforms and not on others. Windows does not
seem to be affected.