Fix memory parsing test
This fixes an issue in this particular test, which resulted in a failure
when loading the full HTML specification because the final memory would
not match the original memory footprint.

This is fixed on the second pass, when memory is already reserved and
re-used instead of allocated. Although this test now takes twice as long
to execute, it produces the correct result.